# Add support for per-module max_methods.