--- a/base/experimental.jl
+++ b/base/experimental.jl
@@ -132,6 +132,7 @@ incomplete knowledge of the argument types.
 Supported values are `1`, `2`, `3`, `4`, and `default` (currently equivalent to `3`).
 """
 macro max_methods(n::Int)
+    0 < n < 5 || error("We must have that `1 <= max_methods <= 4`, but `max_methods = $n`.")
     return Expr(:meta, :max_methods, n)
 end