--- a/src/module.c
+++ b/src/module.c
@@ -1,3 +1,4 @@
+
 // This file is a part of Julia. License is MIT: https://julialang.org/license
 
 /*
@@ -32,6 +33,7 @@ JL_DLLEXPORT jl_module_t *jl_new_module_(jl_sym_t *name, uint8_t default_names)
     m->optlevel = -1;
     m->compile = -1;
     m->infer = -1;
+    m->max_methods = -1;
     JL_MUTEX_INIT(&m->lock);
     htable_new(&m->bindings, 0);
     arraylist_new(&m->usings, 0);
@@ -118,6 +120,21 @@ JL_DLLEXPORT void jl_set_module_infer(jl_module_t *self, int value)
 JL_DLLEXPORT int jl_get_module_infer(jl_module_t *m)
 {
     int value = m->infer;
+    while (value == -1 && m->parent != m && m != jl_base_module) {
+        m = m->parent;
+        value = m->infer;
+    }
+    return value;
+}
+
+JL_DLLEXPORT void jl_set_module_max_methods(jl_module_t *self, int value)
+{
+    self->max_methods = value;
+}
+
+JL_DLLEXPORT int jl_get_module_max_methods(jl_module_t *m)
+{
+    int value = m->max_methods;
     while (value == -1 && m->parent != m && m != jl_base_module) {
         m = m->parent;
         value = m->infer;