Adding
```julia
if isdefined(Base, :Experimental) && isdefined(Base.Experimental, Symbol("@max_methods"))
    @eval Base.Experimental.@max_methods 1
end
```
to `Plots.jl`:
Master:
```julia
julia> using Plots, SnoopCompile, BenchmarkTools

julia> tinf = @snoopi_deep plot(1:10, 1:10)
InferenceTimingNode: 2.085727/12.130569 on Core.Compiler.Timings.ROOT() with 98 direct children

julia> @benchmark plot(1:10, 1:10)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  171.533 μs …   6.345 ms  ┊ GC (min … max): 0.00% … 91.09%
 Time  (median):     177.000 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   187.454 μs ± 173.048 μs  ┊ GC (mean ± σ):  2.97% ±  3.13%

    █▆
  ▂▆██▆▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▁▂▁▂▁▁▁▁▂▁▂▂▁▁▁▁▂▁▁▁▁▁▁▂▁▂▁▁▁▁▁▁▁▁▁▁▁▂▂▃▂ ▂
  172 μs           Histogram: frequency by time          275 μs <

 Memory estimate: 63.12 KiB, allocs estimate: 976.
```
This PR:
```julia
julia> using Plots, SnoopCompile, BenchmarkTools

julia> tinf = @snoopi_deep plot(1:10, 1:10)
InferenceTimingNode: 1.709860/6.303342 on Core.Compiler.Timings.ROOT() with 109 direct children

julia> @benchmark plot(1:10, 1:10)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  226.586 μs …   5.654 ms  ┊ GC (min … max): 0.00% … 92.18%
 Time  (median):     232.976 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   243.641 μs ± 191.944 μs  ┊ GC (mean ± σ):  2.78% ±  3.39%

  ▂▇█▆▅▄▄▃▂▁                                                    ▂
  ███████████▆▅▅▅▅▄▄▄▄▅▄▃▄▄▁▅▁▁▁▄▄▄▁▄▁▁▅▃▄▃▁▁▃▄▁▁▄▄▃▅▃▄▃▃▃▄▅▅▇▇ █
  227 μs        Histogram: log(frequency) by time        378 μs <

 Memory estimate: 63.54 KiB, allocs estimate: 987.
```

Adding `@max_methods 1` to `OrindaryDiffEq`, `DiffEqBase`, `DifferentialEquations`, `VectorizationBase`, `LoopVectorization`, `TriangularSolve`, and `RecursiveFactorization`, I get:
Master:
```julia
julia> using DifferentialEquations, SnoopCompile, BenchmarkTools

julia> lorenz = (du,u,p,t) -> begin
        du[1] = 10.0(u[2]-u[1])
        du[2] = u[1]*(28.0-u[3]) - u[2]
        du[3] = u[1]*u[2] - (8/3)*u[3]
       end
#1 (generic function with 1 method)

julia> u0 = [1.0;0.0;0.0]
3-element Vector{Float64}:
 1.0
 0.0
 0.0

julia> tspan = (0.0,100.0)
(0.0, 100.0)

julia> prob = ODEProblem(lorenz,u0,tspan)
ODEProblem with uType Vector{Float64} and tType Float64. In-place: true
timespan: (0.0, 100.0)
u0: 3-element Vector{Float64}:
 1.0
 0.0
 0.0

julia> alg = Rodas5()
Rodas5{0, true, DefaultLinSolve, Val{:forward}}(DefaultLinSolve(nothing, nothing, nothing))

julia> tinf = @snoopi_deep solve(prob,alg)
InferenceTimingNode: 4.206624/23.152936 on Core.Compiler.Timings.ROOT() with 6 direct children

julia> @benchmark solve($prob, $alg)
BenchmarkTools.Trial: 2299 samples with 1 evaluation.
 Range (min … max):  2.070 ms …  10.632 ms  ┊ GC (min … max): 0.00% … 78.31%
 Time  (median):     2.110 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.171 ms ± 646.150 μs  ┊ GC (mean ± σ):  2.27% ±  6.05%

         ▄▅█▇▅▄▁▁▁
  ▂▃▅▆▆▇▇██████████▆▅▄▃▃▃▃▃▃▃▂▂▂▂▁▂▂▁▁▂▂▃▄▃▄▃▃▃▂▃▃▃▂▂▂▂▂▂▂▂▁▂ ▃
  2.07 ms         Histogram: frequency by time        2.28 ms <

 Memory estimate: 666.94 KiB, allocs estimate: 8955.

julia> lorenz = (du,u,p,t) -> begin
        du[1] = 10.0(u[2]-u[1])
        du[2] = u[1]*(28.0-u[3]) - u[2]
        du[3] = u[1]*u[2] - (8/3)*u[3]
       end
#3 (generic function with 1 method)

julia> prob = ODEProblem(lorenz,u0,tspan)
ODEProblem with uType Vector{Float64} and tType Float64. In-place: true
timespan: (0.0, 100.0)
u0: 3-element Vector{Float64}:
 1.0
 0.0
 0.0

julia> tinf = @snoopi_deep solve(prob,alg)
InferenceTimingNode: 1.233243/2.736248 on Core.Compiler.Timings.ROOT() with 3 direct children
```
This PR:
```julia
julia> using DifferentialEquations, SnoopCompile, BenchmarkTools

julia> lorenz = (du,u,p,t) -> begin
        du[1] = 10.0(u[2]-u[1])
        du[2] = u[1]*(28.0-u[3]) - u[2]
        du[3] = u[1]*u[2] - (8/3)*u[3]
       end
#1 (generic function with 1 method)

julia> u0 = [1.0;0.0;0.0]
3-element Vector{Float64}:
 1.0
 0.0
 0.0

julia> tspan = (0.0,100.0)
(0.0, 100.0)

julia> prob = ODEProblem(lorenz,u0,tspan)
ODEProblem with uType Vector{Float64} and tType Float64. In-place: true
timespan: (0.0, 100.0)
u0: 3-element Vector{Float64}:
 1.0
 0.0
 0.0

julia> alg = Rodas5()
Rodas5{0, true, DefaultLinSolve, Val{:forward}}(DefaultLinSolve(nothing, nothing, nothing))

julia> tinf = @snoopi_deep solve(prob,alg)
InferenceTimingNode: 1.595283/3.539406 on Core.Compiler.Timings.ROOT() with 8 direct children

julia> @benchmark solve($prob, $alg)
BenchmarkTools.Trial: 2025 samples with 1 evaluation.
 Range (min … max):  2.343 ms …  10.539 ms  ┊ GC (min … max): 0.00% … 76.21%
 Time  (median):     2.400 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.465 ms ± 682.168 μs  ┊ GC (mean ± σ):  2.34% ±  6.45%

             ▁▄▅▆█▆▇▆▂▃
  ▂▁▂▂▂▃▃▃▄▅▇██████████▇▇▅▅▅▄▄▃▃▃▃▂▂▃▂▃▂▂▂▂▂▂▃▂▂▃▃▃▃▃▂▂▁▂▂▂▁▂ ▄
  2.34 ms         Histogram: frequency by time        2.54 ms <

 Memory estimate: 717.47 KiB, allocs estimate: 10572.

julia> lorenz = (du,u,p,t) -> begin
        du[1] = 10.0(u[2]-u[1])
        du[2] = u[1]*(28.0-u[3]) - u[2]
        du[3] = u[1]*u[2] - (8/3)*u[3]
       end
#3 (generic function with 1 method)

julia> prob = ODEProblem(lorenz,u0,tspan)
ODEProblem with uType Vector{Float64} and tType Float64. In-place: true
timespan: (0.0, 100.0)
u0: 3-element Vector{Float64}:
 1.0
 0.0
 0.0

julia> tinf = @snoopi_deep solve(prob,alg)
InferenceTimingNode: 1.197684/2.313140 on Core.Compiler.Timings.ROOT() with 3 direct children
```
We'll have to look which type instabilities that appear here are causing the performance regression before adding the option to these packages, but the possible latency improvements here are enticing.