@YingboMa 2021-12-08
What's the implication of `@max_methods 1`? For instance, will `Union{Int,Vector{Int}}` be specialized? What about the iteration protocol?