@KristofferC 2021-12-08
This is what the 3 represents:

```jl
julia> f(x::Int) = 1;

julia> g(x) = f(x)
g (generic function with 1 method)

julia> code_warntype(g, Tuple{Any})
MethodInstance for g(::Any)
  from g(x) in Main at REPL[2]:1
Arguments
  #self#::Core.Const(g)
  x::Any
Body::Int64
1 ─ %1 = Main.f(x)::Core.Const(1)
└──      return %1


julia> f(x::Float64) = 2.0;

julia> code_warntype(g, Tuple{Any})
MethodInstance for g(::Any)
  from g(x) in Main at REPL[2]:1
Arguments
  #self#::Core.Const(g)
  x::Any
Body::Union{Float64, Int64}
1 ─ %1 = Main.f(x)::Union{Float64, Int64}
└──      return %1


julia> f(x::Float32) = 3.0f0;

julia> code_warntype(g, Tuple{Any})
MethodInstance for g(::Any)
  from g(x) in Main at REPL[2]:1
Arguments
  #self#::Core.Const(g)
  x::Any
Body::Union{Float32, Float64, Int64}
1 ─ %1 = Main.f(x)::Union{Float32, Float64, Int64}
└──      return %1


julia> f(x::String) = "4"
f (generic function with 4 methods)

julia> code_warntype(g, Tuple{Any})
MethodInstance for g(::Any)
  from g(x) in Main at REPL[2]:1
Arguments
  #self#::Core.Const(g)
  x::Any
Body::Any
1 ─ %1 = Main.f(x)::Any
└──      return %1
```