@chriselrod 2021-12-08
> What's the implication of `@max_methods 1`? For instance, will `Union{Int,Vector{Int}}` be specialized? What about the iteration protocol?

It will not cause problems for/is not applicable to the `iteration` protocol. This does not rely on method matching:
```julia
valuestate = iterate(iter, state)
valuestate === nothing || break
value, state = valuestate
```
Union splits should (likewise) generally be fine.