@KristofferC 2023-01-16
Shouldn't this have any tests?