@chriselrod 2023-01-16
What should tests look like?
If you want an example use
https://github.com/JuliaPlots/Plots.jl/blob/ce961acec433d34be00470e2c07500e9be894080/src/Plots.jl#L6-L8

I should probably add it to LoopVectorization, too.