@KristofferC 2023-01-16
> Test like checking that the functionality works.

Like using the functionality `@max_methods` and checking that it had the intended effect, like checking that inference stops after it reaches the max methods instead of the default 3 etc?