@KristofferC 2023-01-17
Okay, this PR did less than I thought and I didn't find the tests of the original `max_methods` PR due to https://github.com/JuliaLang/julia/pull/48316